requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=8.0.0
cairosvg>=2.5.0
mysql-connector-python>=8.0.0
